# Conv1d optimize