--- a/candle-core/src/cpu_backend.rs
+++ b/candle-core/src/cpu_backend.rs
@@ -1040,25 +1040,25 @@ impl<'a> Map2 for Conv1D<'a> {
         let dst_elems = p.c_out * l_out * p.b_size;
         let mut dst = vec![T::zero(); dst_elems];
         // The output shape is [b_size, c_out, l_out]
-        for b_idx in 0..p.b_size {
-            let inp_idx = b_idx * inp_s0;
-            let dst_idx = b_idx * p.c_out * l_out;
-            for dst_c_idx in 0..p.c_out {
-                let dst_idx = dst_idx + dst_c_idx * l_out;
+        for offset in 0..p.k_size {
+            for b_idx in 0..p.b_size {
+                let inp_idx = b_idx * inp_s0;
+                let dst_idx = b_idx * p.c_out * l_out;
                 for dst_l in 0..l_out {
                     let dst_idx = dst_idx + dst_l;
-                    let mut d = T::zero();
-                    for offset in 0..p.k_size {
+                    for dst_c_idx in 0..p.c_out {
+                        let dst_idx = dst_idx + dst_c_idx * l_out;
                         let src_l = (p.stride * dst_l + offset)
                             .saturating_sub(p.padding)
                             .min(p.l_in - 1);
+                        let mut d = T::zero();
                         for src_c_idx in 0..p.c_in {
                             let inp_idx = inp_idx + src_c_idx * inp_s1 + src_l * inp_s2;
                             let k_idx = dst_c_idx * k_s0 + src_c_idx * k_s1 + offset * k_s2;
                             d += inp[inp_idx] * k[k_idx]
                         }
+                        dst[dst_idx] += d
                     }
-                    dst[dst_idx] = d
                 }
             }
         }