--- a/candle-core/src/lib.rs
+++ b/candle-core/src/lib.rs
@@ -40,6 +40,7 @@ pub mod backprop;
 mod conv;
 mod convert;
 pub mod cpu_backend;
+pub mod cpu_kernels;
 #[cfg(feature = "cuda")]
 pub mod cuda_backend;
 mod device;