This optimizes the cpu implementation for `conv1d`: it changes the loop order so that the operation result in a large dot product and `ggblas` is used to implement this in an efficient way using simd where available. This does not use multiple threads. It is less efficient memory-wise than the naive implementation as the input tensor is made contiguous.
Benchmarking results on my macbook (m2pro 2022).
- On the conv1d benchmark in `candle-core/examples`, the runtime goes from 1.1s to 101ms.
- On whisper, the two first convolutions go from 226ms + 568ms to 30ms + 65ms.